new hierarchy

+++ headers/best_worst.h
@@ -1,5 +1,5 @@
 #ifndef BESTWORST
 #define BESTWORST
-void determineBestMonarch(int ** population, double ** costTable, double ** delayTable, int num_cols, int bin, int ** bestMonarch, int num_rows);
+void determineBestMonarch(int ** population, double ** costTable, double ** delayTable, int num_cols, float bin, int ** bestMonarch, int num_rows);
 int determineWorstMonarch(int ** Population, double ** costTable, double ** delayTable, int num_cols, int num_rows);
 #endif
